style: 테스트 서비스 파일명 수정
