+++ src/main/java/com/spring/moji/controller/TestController.java
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 import com.spring.moji.dto.request.TestRequestDTO;
 import com.spring.moji.dto.response.TestResponseDTO;
-import com.spring.moji.service.TestService;
+import com.spring.moji.service.TestServiceImpl;
 import com.spring.moji.util.S3Util;
 import lombok.RequiredArgsConstructor;
 
@@ -23,7 +23,7 @@ import lombok.RequiredArgsConstructor;
 public class TestController {
 	@Value("${kakao.api-key}")
 	private String kakaoApiKey;
-	private final TestService testService;
+	private final TestServiceImpl testService;
 	private final S3Util s3Util;
 
 	@GetMapping("/diary/cover")
